Updated shop.lua to handle multiple mods using the same instance of the library
Previous iteration of shop.lua could not handle several mods using the same file.
The load function has been removed, so mods don't have to manually load it anymore.

A bug has been introduced by doing this. Currently, the option for every weapon added will be enabled regardless of what the option is toggled to. This will be fixed when a solution to the issue can be found.