Changed all occurances of AddPawn to AddPawnName
Calls to `AddPawn` for each pawn added by each mod, was added in a previous commit, in order to add the pawn names to lists that let the game know which pawns has been added.

`AddPawn` Not only adds the pawn names to this list. It also creates pawn objects with the table provided, based on the default pawn. This meant that any previous base object that was used, would now be forgotten.
For example if Roach2 was based on Roach1. Using AddPawn for Roach2 would make Roach2 forget about its base Roach1.

Each call has been replaced by calls to `AddPawnName`. A function which only adds the pawn to the pawn list used by the game, and nothing else.