Library trait.lua: Changed to update all traits whenever the save data is updated
Pilot traits would not update when entering Test Mech Scenario, because the save data at that time was not up to date at the time of entering.
Instead of updating traits when entering Test Mech Scenario and when loading the game; update them when the save data is updated instead.

This appears to be a much more reliable way to keep traits updated.
An added bonus is that many conditional traits could become easier to make, as many actions trigger a save, which in turn will update traits based on new conditions.