WeaponPreview: Removed need for mod id
When using the library to display animation (and emitters), the animations are separated into multiple 1 frame animations. Previously the mod id of the mod using the library would be used as an identifier for these animations.
This update changes it to a global naming convention instead. This removes the need to tie an instance of the library to a specific mod.

Later commits will ensure every mod using the library will use the same global naming convention, even if they don't run the same version.