WeaponPreview: Add checks to GetTargetArea and GetSkillEffect
The vanilla weapon `Support_Missiles` calls `GetTargetArea` with `nil` as the first argument. It looks like a typo, but it doesn't affect the weapon itself since it doesn't use the argument for anything.

WeaponPreview hooks into every weapons' `GetTargetArea`, and would throw an error when the first argument is `nil`.

This has now been fixed.