Into The Wild: Update bonus objective Specimen
At the end of a mission, enemies are briefly changed to neutral units to prevent them from retreating. This makes the objective appear failed during an `onMissionEnd` event.

This has now been accounted for by storing the number of enemies on the board before converting them, and referring to this cached value while they are neutral.

However, if any other mod would attempt to count the enemies in an `onMissionEnd` event, they would not find them.

Potential solution could be to recode `missionEnd` to allow for enemies to be tagged as not retreating instead of changing their team. This would probably be best added to the mod loader though.