Dune Striders: Update utils.lua
Change utils.GetGenericImpactSoundScript to utils.GetImpactSound

Now returns a string for the appropriate impact sound for a tile, taking into account any unit there with an impact material; instead of returning a script with code to play the sound.
Any code broken by this change will be addressed in a later commit.